Move task state into a 'Tasks' subdir
Having CuratorStateStore scan the root directly causes it to
misinterpret data which was put there by other sources, eg
CuratorConfigStore's [fwkname]/Configurations/[confid]/.

To CuratorStateStore, this path would look like an execName of
"Configurations" and a taskName of "[confid]".

This also allows us to remove CuratorStateStore's check to avoid
listing the FrameworkID file as an executor name.