Fix test flakes
1. The DefaultSchedulerTest had a race.  It's possible Offers are
accepted, and the test validates their contents and then tries to
validate Block state before the Block has been called back alerting it
to the success or failure of it's launch.

2. Hard to say why the HealthCheck testSuccess was failing.  There was a
shared threadpool for all tests with a single thread.  Perhaps this was
causing trouble as it dealth with handling failed threads and starting
new ones.  In practice the HealthCheck threadpool should be
substantially larger than 1, and probably dynamically sized so each
health-check has it's own Thread.  I have given each test it's own
single threaded thread pool now.