Implement fixed and dynamic Placement Constraints
Provides a bunch of built-in placement rule implementations.

- PlacementRule: A static rule which accepts/declines Offers for the
  purpose of launching a single task.
- PlacementRuleGenerator: A dynamic rule generator which creates a
  PlacementRule for the next task to be launched, based on current
  task deployment.

Note that PlacementRules may internally be composites of other
PlacementRules, via AndRule/OrRule/NotRule (which contrary to their
names are actually performing set logic on returned Offer Resources).

The provided PlacementRuleGenerator implementations support things
like 'avoid (or colocate with) any tasks with this type', and 'don't
launch more than N tasks of this type on agents with this attribute'.

Summary of changes to existing dcos-commons:
- Store a string representation of the offer attributes, if any, as
  a label in the TaskInfo. The string representation matches the spec
  provided by Mesos for attribute strings.
- Replace existing avoidAgents/colocateAgents logic with a rules-based
  implementation.
- Replace "taskType + '-' + index" in DefaultTaskSpecification with a
  a new 'toTaskName()' utility function in TaskUtils.