Implement Placement Constraints
This pass includes most of the dcos-commons integration, replacing
the previous hand-coded colocation/avoidance constraints. A following
commit will expand on this by adding additional implementations of
PlacementRules and PlacementRuleGenerators, but the dcos-commons
plumbing should all be included here.

The most notable change is the additional data being serialized into
the TaskInfo, namely the Task type name (from TaskSet), and the
Offer Attributes, if any, that were present in the accepted offer
upon task launch. Strictly speaking these values will only be used
by PlacementRuleGenerators that aren't included in this PR (e.g.
one which limits the number of instances of a given type running
against a given attribute) but I regardless wanted to include all
the plumbing up-front in this PR.