Support atomically storing multiple values in CuratorPersister
This allows atomic operations such as:
- Create and/or update multiple TaskInfos at once
- Create/update a TaskInfo+TaskStatus at once

Curator transactions require foreknowledge of whether nodes need to be
created or merely modified, as there is no in-transaction statement
for 'create or update'. In addition, any parent nodes need to be
created manually.

The logic here is a workaround for these deficiencies. The idea is to
compose a transaction based on the current state of the ZK tree which
creates nodes that don't exist and updates nodes that do exist, then
the composed transaction is executed.

This composition/execution cycle is attempted a maximum of three
times, to handle potential situations where the tree is modified
out-of-band by another party between the composition step and the
execution step. In practice this shouldn't come up but it's nice to
have.