+minor addition for dynamic ports and VIP setup.
Made modifications in ResourceRequirements. Enabled passing any resource value as an environment variable to the application. EnvName is sync'ed to the reservation labels.
Modifications in OfferEvaluation as well.

specific problems: dynamic ports and VIP setup.

If port value is 0, it is processed differently in the Offer evaluation. A port value is assigned and ResourceRequirement is updated.

VIP setup is done by providing a label. Since port value can be empty, VIP setup (adding to TaskInfo Discovery label) is delayed till and offer is found - fulfilledRequirement.

While returning from Offer evaluate I still use Resource, not ResourceRequiement. This does unnecessary conversion. Intentionally done like that for the moment because Resource object are passed in general in the cod. This helps to make sure Resource labels are properly populated, so conversion can happen with no information change.

ResourceRequirement add to the reservation labels with a predefined key (to save the bookkeeping information). If label key (i.e. ENV_KEY or VIP_KEY) exits, it overwrites them.