api-port is dynamically set by Marathon.  In yml file, it should be “api-port: {{PORT0}}”
Marathon dynamically selects a port, and pass this info to the framework via env variable, so framework can start the service on that selected port.

Somehow in svc.yml, we set a port number 8080, which is not reserved. Therefore, framework api service would not work.

See Marathon mustache configuration for more info:

hello-world:
…
 "DCOS_MIGRATION_API_PATH": "/v1/plan",
    "MARATHON_SINGLE_INSTANCE_APP":"true",
    "DCOS_SERVICE_NAME": "{{service.name}}",
    "DCOS_SERVICE_PORT_INDEX": "0",
    "DCOS_SERVICE_SCHEME": "http”
…
  "portDefinitions": [
    {
      "port": 0,
      "protocol": "tcp",
      "name": "api",
…

cassandra:
…
  "MARATHON_SINGLE_INSTANCE_APP":"true",
  "DCOS_SERVICE_NAME": "{{service.name}}",
  "DCOS_SERVICE_PORT_INDEX": "0",
  "DCOS_SERVICE_SCHEME": "http"
…
"portDefinitions": [
    {
      "port": {{service.api_port}},
      "protocol": "tcp",
      "name": "api",
…
,"API_PORT":"{{service.api_port}}"
…