YAML: Add 'colocate-types'/'avoid-types', remove 'placement'
In practice, type colocate/avoid is the most common thing needed in
the yaml specs so far, so this optimizes for that case. Devs may still
implement their own custom placement rule trees, but that now involves
doing it in Java. Authoring placement rule trees as yaml is
tricky/error-prone.

YAML:
- avoid-tasks/colocate-tasks are added to pod spec
- custom 'placement' option is removed. devs who need custom placement
  logic are likely to be using java anyway.
Java:
- custom placement option is retained, allowing java devs to implement
  their own placement logic, possibly even extending to custom
  PlacementRule implementations