cli: remove plan active (#457)
* - Plan and Phase return WAITING, if interrupted and not completed.
- IN_PROGRESS status only for plans and phases

DefaultPlan and DefaultPhase used to check if interrupted in getStatus. That is omitted when we remove Element.setStatus. Putting back this feature in PlanUtils.getStatus (only Plan and Phase call this)!

* use proper docker protobugs and fix rlimits NPE (#441)

* - Plan and Phase return WAITING, if interrupted and not completed.
- IN_PROGRESS status only for plans and phases

DefaultPlan and DefaultPhase used to check if interrupted in getStatus. That is omitted when we remove Element.setStatus. Putting back this feature in PlanUtils.getStatus (only Plan and Phase call this)!

* use proper docker protobugs and fix rlimits NPE (#441)

* use proper docker protobugs and fix rlimits NPE (#441)

* rebase master

    - Plan and Phase return WAITING, if interrupted and not completed.
    - IN_PROGRESS status only for plans and phases

* remove v1/plan/status (does not exist anymore)