Faster service rebuilds via caching of go artifacts
Adds a 'build_go_exe.sh' script which implements caching logic to
avoid unnecessary rebuilds of go executables. This is done by
comparing the SHA1 of a native/local build (faster than cross-builds)
to detect changes in the go build artifact.

This script is now used for building per-service CLI executables as
well as the common `bootstrap` executable.

As a result, rebuilding a service goes from ~40s to ~5s.