INFINITY-1876 Specify minDcosVersion in package, not release script
The current minDcosVersion handling was only in the release script because our stub-universe.zips were stuck on universe-2.x, which didn't understand this field and which would refuse to install packages that contained it. So at the time we omitted it from stub-universe.zips and just added it when releasing them to universe.

Now that we're on universe-3.x packages in the form of stub-universe.jsons, we can specify it explicitly in the source-controlled package files. This PR also updates the release tooling to require that it already be populated, as it should currently be AT LEAST 1.8 for SDK-based services (leaving the field unset means the package can be installed back to 1.6 IIRC).

TODO separately: Remove the option/envvar from the jenkins release job as this makes the option a noop.