Patch for PersistentLaunchRecorderTest (#1169)
* Patch for PersistentLaunchRecorderTest

This test would only fail when ordering was changed and would also
segfault the JVM. The segfault was due to an incompatibility with
gradle's native jansi module. This was fixed with explicitly importing
a newer module. The initial failure was due to
PeristentLaunchRecorderTest shadowing the beforeClass method of parent
DefaultCapabilities test suite. Parent classes were made final to avoid
future NPEs in other classes.

* Patch DefaultCapabilities instead

* Remove unwanted comment