Update to allow tasks to choose other agents if there are no persistent resources
Even if a task expects reserved resources, if the resources were not persistent
on that node, it should be ok to try and start the task on some other agent.
This allows tasks to failover to other agents in case an agent dies or
disappears.

Also, if we didn't get the expected resource from an offer use new Resource Ids.
The old resources will be free'd by us.