Skip task log collection in soak environment.
We use the continuous integration py.test setup in soak to run some tests.
When these fail in continuous integration, we collect all the cluster logs,
since they're all likely relevant to the failure.

In a soak cluster, however, there are a large (~20) frameworks and marathon
apps running, as well as many metronome jobs etc, most of which are not even
related to infinity, let alone the failure at hand.  Collecting all these
logs causes cascading failures with test failures taking longer than the
periodicity of the test execution.