Fix grep matching and subshell output expansion.
`grep [${aws_profile}]` was matching a bracket regular expression instead of a
string with literal brackets.

Also the expansion of `$( echo "$PROFILES" | grep "\[${aws_profile}\]" )`
needs to be enclosed in double quotes so that the != operator works and
doesn't give us "unary operator expected" errors when the output is empty.