INFINITY-3107 Trim 'beta-' prefix for default service name in CLI (#2217)
If `--name` (or the `<modname>.service_name` option) isn't specified, the CLI currently defaults to using the package name. However, this doesn't work for `beta-foo` packages, where the default service name is `foo`.

With this change, the CLI will align with the service name convention used by beta packages. Users interacting with default installs of beta packages will no longer be required to provide the `--name` argument. If they provide a custom name then they will need to specify the `--name` argument, as they would with a GA package.