Change of tack: Only store resource label when running a namespaced service
Adding a resource label breaks upgrade compatibility for existing services, unless a lot of complexity is added. Mesos makes this pretty much impossible to do properly.
So instead, let's only include the service name label if the service is configured with a namespace. This effectively disables this label for existing services, avoiding the compatibility issues.