INFINITY-3500 Upstream default service state manager (#2468)
Removes the previous requirement that the developer implement their own tracking of which services are running in a multi-service scheduler. Instead, the developer only needs to implement a callback method which consistently constructs (or reconstructs) the service from context data that the developer controls. The "context data" (max 100KB) allows the developer to store any configuration/context needed to reconstruct the service by their own callback.

Switches the example `hello-world` implementation to use this new mechanism.