DCOS-38847 Use autoip instead of VIP for fetching config artifacts
With this PR, config fetching is switched from a VIP at port 80 to an autoip endpoint directly on the dynamic port.

Usage of VIPs was originally introduced before autoip was an option. VIPs have the issue of always running on the same port 80: Customers may have implemented firewall rules on agents which effectively block this access.