DCOS-39121 Automatically renew creds with 'maws' when upload fails
If an upload to S3 fails, an attempt will be automatically made to reauth before reattempting the upload. If reauth fails, or if `maws` is not present in the local `PATH`, then it reverts to current behavior and just gives up.

Automatically determines the profile to renew via the `AWS_PROFILE` envvar, with a fallback of extracting the profile name from the credentials INI file.