DCOS-39389 Support customizing marathon name and VIP TLD (#2592)
Allow the operator to manually specify a custom name for marathon. This is used when creating URLs which point back to the scheduler, and may need to be customized if the scheduler is running in a MoM.

Similarly, we currently allow manual customization of the autoip TLD, but didn't yet support the same for the VIP TLD. Add support for that.