[DCOS-29704] Track plan statuses in metrics (#2538)
This is a WIP (no tests yet) PR to:
1. Start tracking the ordinal of the plan status in a metric per plan (namespaced for multi-scheduler).
2. Add a simple timed executor that scrapes plan status.

@nickbp: I put the `PlanReporter` instantiation into `DefaultPlanCoordinator` as it seemed to be the best single leverage point for access to all plan managers.

I went with tracking the ordinal as it is relatively simple and creates just one metric per plan. Unfortunately, tags are not supported by the metrics library as of yet so I can't just have one metric with many tags. I'm also not sure that having one metric with many tags would give the desired result anyways, which is to have a consistent status of when plan status changes occur.

Also, I hardcoded the interval to 5 seconds. It could realistically go higher or lower.