DCOS-42539 (backport) Improve recovery config ID selection (#2694)
Previously, we were selecting a random task in the pod when performing a recovery (restart/replace) operation. This interim fix improves on the existing (lack of) selection logic by making it do the right thing in today's typical cases. A larger change will be made in sdk-0.50+ to improve the config update flow itself to avoid mixed configs in the same pod unless explicitly needed via e.g. a custom plan.

- Only examine tasks which are being included as part of the recovery operation, rather than all tasks in the pod.
- Among tasks being recovered, prefer the tasks which have a RUNNING goal state, over those with a FINISHED or ONCE goal state.
- Improve logging of the decision itself.

... and while we're in sdk-0.40, this PR also fixes an encoding issue in Kafka's config options for DCOS-42473