DCOS-43056: Suppress when plans are COMPLETE, ERROR, or WAITING (#2713)
At the moment only COMPLETE was the only state that was considered to be "not doing work" and therefore eligible for a SUPPRESS call. With this change, plans that are in WAITING or ERROR states are also considered acceptable when deciding to SUPPRESS.

Came up in a service with a custom plan which defaulted in a WAITING state. Even though the `deploy` plan was complete, the custom plan in WAITING was causing the service to not suppress.