DCOS-42539 Break up statestore updates into separate object (#2708)
This PR is an up-front internal refactoring that shouldn't change actual SDK behavior. A separate follow-up PR will change behavior to ensure that ONCE/FINISHED tasks are correctly updated during config changes.

Previously, the `LaunchOfferRecommendation` produced by offer evaluation had two modes based on a `shouldLaunch()` value:
- launch task+update StateStore
- just update StateStore

This change breaks out the "update StateStore" part of `LaunchOfferRecommendation` into a dedicated `StoreTaskInfoRecommendation`, and updates the `OfferRecommendation` interface to have `getOperation()` return an `Optional<Operation>`. This allows for cases like `StoreTaskInfoRecommendation` where there's nothing to be sent to Mesos when the recommendation is accepted. This should make it easier to have other internal bookkeeping tasks like this in the future.