[DIAGNOSTICS] Fix condition where scheduler task can't be found.
Due to a name mismatch. The DCOS_SERVICE_NAME label in SDK services is usually
prefixed by a slash, while the actual DCOS "service_name" outputted by `dcos
service --json` may or may not be prefixed by the slash.

This change make sure to get the scheduler task with and without a slash prefix.