[DCOS-47829] Bump cli to version 0.7.8 (#2920)
Downloaded the CLI from https://downloads.dcos.io/binaries/cli/linux/x86-64/dcos-1.12/dcos

We were using a very old version of CLI, this PR updates the version of CLI being used. Verified that this CLI works with `1.10` and `1.12` and master. (it prints a warning but it works nonetheless)

We can/SHOULD revert this in case of any unforeseen circumstances or nightlies start to break.