Don't escape characters for custom YAML.
As per the manual/spec[1] on which the Java library we use bases its
implementation, one needs to use triple braces for variables that
should not have escaped characters when expanded. Without this, for
example, if one has quotes in the YAML content, they will get expanded
to their HTML entity: `&#34;`.

[1]: http://mustache.github.io/mustache.5.html