[COPS-4469] Shutdown the scheduler when service user changes in a ServiceSpec update
Previously the SDK would detect a change of users in the ServiceSpec, emit a warning
and *not* propagate and store these updates into ZK. The scheduler would however
continue to keep running with the updated user and this can cause issues with downstream
components such as volumes which may have permissions set for the previous user.

This change explictly disallows changing of users on a previously deployed service. The
scheduler issues a *shutdown* with a prescriptive message of how to resolve this issue.

Operators should note the scheduler crash-looping with the following message in the scheduler logs
when a service user change is applied:

    'INVALID CONFIGURATION UPDATE. Cannot change user of deployed service from 'userA' to 'userB'.
    Revert to previous user 'userA' to proceed. Current set of configuration updates will NOT be applied.'

The violating updates to the ServiceSpec are again *not* propagated to ZK maintaining existing
semantics.