[DCOS-49331] Write data to zk only when needed (#2946)
When SDK is writing data to zk, in some cases it is creating a `KeeperException.NodeExistsException` and this may lead to large exhibitor log files in unstable (agent disconnected etc..) clusters with multiple SDK schedulers running. This PR addresses this by problem checking for the existing zk node data before creating/updating it.