[DCOS-52239] Replace the spec section with a quickstart one
Remove the short section on the service spec - there is a very similar
but better one in the developer guide already, and it's not really
appropriate to attack a complete newcomer with such details.

Instead, re-add the quick-start section that should be more useful to
someone who wants to start using the SDK. In there, point them at the
dcosdev tool and the helloworld example for the lack of a better HOWTO
guide.