[DCOS-53557][COPS-4781] Expose Kibana web server through an nginx proxy.
This was done to address the issue surfaced by COPS-4781: accessing the Kibana
service URL resulted in a redirect to an invalid URL. Namely:

GET /service/kibana/ => /service/kibana/service/kibana/app/kibana

The expected redirect should have been:

GET /service/kibana/ => /service/kibana/app/kibana

This happened because of a breaking change in Kibana 6.x which caused requests
to the Kibana web server hitting the root path ("/") resulted in an invalid
redirect, depicted above.

This could be the place where this bug was introduced:
https://github.com/elastic/kibana/pull/21408/files#diff-230bb53cded88451213fc31d8a4b0774L103

The solution here was to add an nginx proxy in between Admin Router and the
Kibana web server, that rewrites requests that would hit the Kibana web server
on the root path to hit it on "/app/kibana" instead.

I also took the opportunity to extract the Kibana Marathon app command to an
actual bash script, which is much more maintainable than the previous one-line
"cmd" directive.