[DCOS-61858] Quota Framework Uninstall Fixes (#3215)
- Background

This series of patches fixes the uninstall bug found in DCOS-61858 where frameworks under the same group & quota'ed role where sharing a service-principal. Upon uninstallation, the framework being uninstalled DESTROYED and UNRESERVED resources belonging to other frameworks in the same role.

This bug occurs on DC/OS 2.0 and later clusters where the service-principal is shared.

- Resolution

This series of patches applies the framework_id label to all resources created by the framework.
The framework scheduler now filters out any offer with resources that have a mismatched framework_id label.

The bulk of this PR is propagating the framework_id from where the scheduler subscribes to Mesos, to where OfferEvaluation and the resources are built.

This patch-set also adds a debug reservations endpoint (DCOS-44705) for easier debugging of which reservations are owned by the current set of running tasks.