fix COPS-5547
customers might not have `REMOTE_SEEDS`. in that case the comma in the end tells the config parser that the list end with a NULL, which in turn is converted to `127.0.0.1`, which then results in the debug log constantly spewing errors about the gossip protocol not being able to communicate with localhost.

closes COPS-5547